requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-environ==0.8.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==12.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
